Perform Hierarchical Clustering on Methylation Data

Produces a hierarchical clustering tree based on the input matrices of counts.

Usage

fuse.cluster(K0, K1)

Arguments

K0

Integer or numeric matrix with unmethylated counts

K1

Integer or numeric matrix with methylated counts

Value

Clustering tree as a matrix, with the same structure as hclust, including the columns

m1

ID of first data point. <0 is original point, >0 is row of previous merge.

m2

ID of second data point. <0 is original point, >0 is row of previous merge.

logl_tot

Change in total log-likelihood of merge.

logl_merge

Total cost of points in this merge.

genomic_dist

Penalty for genomic distance between points.

Examples

# Example: Clustering generated data
set.seed(1234)
K0 <- matrix(
  rep(c(sample(0:20, 200, replace = TRUE), sample(20:40, 200, replace = TRUE)), 2),
  nrow = 100, byrow = TRUE
)
K1 <- matrix(
  rep(c(sample(20:40, 200, replace = TRUE), sample(0:20, 200, replace = TRUE)), 2),
  nrow = 100, byrow = TRUE
)
tree <- fuse.cluster(K0, K1)
tree
#>         m1   m2    logl_tot  logl_merge genomic_dist
#>   [1,] -40  -41    4.745890    4.745890      1.14473
#>   [2,] -90  -91    4.745890    4.745890      1.14473
#>   [3,]  -6   -7    7.394532    7.394532      1.14473
#>   [4,] -56  -57    7.394532    7.394532      1.14473
#>   [5,] -13  -14    8.257038    8.257038      1.14473
#>   [6,] -63  -64    8.257038    8.257038      1.14473
#>   [7,] -44  -45    8.531066    8.531066      1.14473
#>   [8,] -94  -95    8.531066    8.531066      1.14473
#>   [9,] -42  -43   10.069508   10.069508      1.14473
#>  [10,] -92  -93   10.069508   10.069508      1.14473
#>  [11,]  -3   -4   10.074838   10.074838      1.14473
#>  [12,] -53  -54   10.074838   10.074838      1.14473
#>  [13,] -48  -49   10.287519   10.287519      1.14473
#>  [14,] -98  -99   10.287519   10.287519      1.14473
#>  [15,] -19  -20   10.810084   10.810084      1.14473
#>  [16,] -69  -70   10.810084   10.810084      1.14473
#>  [17,] -31  -32   11.111468   11.111468      1.14473
#>  [18,] -81  -82   11.111468   11.111468      1.14473
#>  [19,] -16  -17   11.374850   11.374850      1.14473
#>  [20,] -66  -67   11.374850   11.374850      1.14473
#>  [21,] -38  -39   11.488309   11.488309      1.14473
#>  [22,] -88  -89   11.488309   11.488309      1.14473
#>  [23,] -28  -29   14.688620   14.688620      1.14473
#>  [24,] -78  -79   14.688620   14.688620      1.14473
#>  [25,] -26  -27   15.170902   15.170902      1.14473
#>  [26,] -76  -77   15.170902   15.170902      1.14473
#>  [27,] -35  -36   17.209725   17.209725      1.14473
#>  [28,] -85  -86   17.209725   17.209725      1.14473
#>  [29,]  -9  -10   18.407134   18.407134      1.14473
#>  [30,] -59  -60   18.407134   18.407134      1.14473
#>  [31,] -21  -22   19.464621   19.464621      1.14473
#>  [32,] -71  -72   19.464621   19.464621      1.14473
#>  [33,]  -1   -2   19.722421   19.722421      1.14473
#>  [34,] -51  -52   19.722421   19.722421      1.14473
#>  [35,] -30   17    9.753578   20.865046      1.14473
#>  [36,] -80   18    9.753578   20.865046      1.14473
#>  [37,] -11  -12   22.482809   22.482809      1.14473
#>  [38,] -61  -62   22.482809   22.482809      1.14473
#>  [39,]  -5    3   15.312638   22.707169      1.14473
#>  [40,] -55    4   15.312638   22.707169      1.14473
#>  [41,]   7  -46   15.542385   24.073451      1.14473
#>  [42,]   8  -96   15.542385   24.073451      1.14473
#>  [43,]  21    1    9.375008   25.609206      1.14473
#>  [44,]  22    2    9.375008   25.609206      1.14473
#>  [45,]  13  -50   18.975164   29.262683      1.14473
#>  [46,]  14 -100   18.975164   29.262683      1.14473
#>  [47,] -23  -24   31.299302   31.299302      1.14473
#>  [48,] -73  -74   31.299302   31.299302      1.14473
#>  [49,] -34   27   15.656186   32.865910      1.14473
#>  [50,] -84   28   15.656186   32.865910      1.14473
#>  [51,]  -8   29   15.405146   33.812280      1.14473
#>  [52,] -58   30   15.405146   33.812280      1.14473
#>  [53,]   5  -15   25.819240   34.076278      1.14473
#>  [54,]   6  -65   25.819240   34.076278      1.14473
#>  [55,]  19  -18   23.494407   34.869257      1.14473
#>  [56,]  20  -68   23.494407   34.869257      1.14473
#>  [57,]  35  -33   26.309405   47.174451      1.14473
#>  [58,]  36  -83   26.309405   47.174451      1.14473
#>  [59,]  33   11   18.344049   48.141308      1.14473
#>  [60,]  34   12   18.344049   48.141308      1.14473
#>  [61,]  15   31   18.497838   48.772544      1.14473
#>  [62,]  16   32   18.497838   48.772544      1.14473
#>  [63,]  25   23   20.925483   50.785005      1.14473
#>  [64,]  26   24   20.925483   50.785005      1.14473
#>  [65,] -37   43   26.874343   52.483549      1.14473
#>  [66,] -87   44   26.874343   52.483549      1.14473
#>  [67,]  41  -47   29.145856   53.219307      1.14473
#>  [68,]  42  -97   29.145856   53.219307      1.14473
#>  [69,]  47  -25   27.985280   59.284581      1.14473
#>  [70,]  48  -75   27.985280   59.284581      1.14473
#>  [71,]  39   51   12.439239   68.958688      1.14473
#>  [72,]  40   52   12.439239   68.958688      1.14473
#>  [73,]  37   53   14.293104   70.852191      1.14473
#>  [74,]  38   54   14.293104   70.852191      1.14473
#>  [75,]  65    9   25.896198   88.449255      1.14473
#>  [76,]  66   10   25.896198   88.449255      1.14473
#>  [77,]  55   61    6.795741   90.437542      1.14473
#>  [78,]  56   62    6.795741   90.437542      1.14473
#>  [79,]  67   45   11.600742   94.082732      1.14473
#>  [80,]  68   46   11.600742   94.082732      1.14473
#>  [81,]  57   49   30.253565  110.293926      1.14473
#>  [82,]  58   50   30.253565  110.293926      1.14473
#>  [83,]  59   71    9.290299  126.390296      1.14473
#>  [84,]  60   72    9.290299  126.390296      1.14473
#>  [85,]  77   69    5.829966  155.552089      1.14473
#>  [86,]  78   70    5.829966  155.552089      1.14473
#>  [87,]  63   81   26.686063  187.764994      1.14473
#>  [88,]  64   82   26.686063  187.764994      1.14473
#>  [89,]  83   73    6.400751  203.643238      1.14473
#>  [90,]  84   74    6.400751  203.643238      1.14473
#>  [91,]  75   79   24.054433  206.586420      1.14473
#>  [92,]  76   80   24.054433  206.586420      1.14473
#>  [93,]  89   85   21.591884  380.787211      1.14473
#>  [94,]  90   86   21.591884  380.787211      1.14473
#>  [95,]  87   91   16.117039  410.468452      1.14473
#>  [96,]  88   92   16.117039  410.468452      1.14473
#>  [97,]  93   95 2255.945351 3047.201014      1.14473
#>  [98,]  94   96 2255.945351 3047.201014      1.14473
#>  [99,]  97   98    0.000000 6094.402028      1.14473